+++ backend/src/main/java/heartbeat/controller/report/dto/response/MeanTimeToRecoveryOfPipeline.java
@@ -1,6 +1,7 @@
 package heartbeat.controller.report.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -18,6 +19,6 @@ public class MeanTimeToRecoveryOfPipeline {
 	@JsonProperty("step")
 	private String pipelineStep;
 
-	private double timeToRecovery;
+	private BigDecimal timeToRecovery;
 
 }
